fix: use int64_t/int32_t in ld, lw, sd, sw

+++ test/unit/insns/ut_lw.cpp
@@ -7,11 +7,11 @@ TEST_F(ut_insns, decode_and_execute_rv64i_lw) {
     // 0x00c92503 : lw a0, 12(s2)
     insts.push_back(0x00c92503);
     
-    uint64_t *addr = (uint64_t *)(GetStackPointer() + 12);
+    int32_t *addr = (int32_t *)(GetStackPointer() + 12);
     *addr = 0x10001000;
 
     auto regr = reg::a0;
-    uint64_t result = 0;
+    int32_t result = 0;
     SetIReg(reg::s2, GetStackPointer());
     ExecuateInst();
     result = GetIReg(regr);
